make TOYNET, breakup my_diagram.py into logical groupings
