+++ clean_toynet.py
@@ -1,0 +1,3 @@
+from mininet.clean import cleanup
+
+cleanup()
