+++ toydiagram/resources/host/host.png
