+++ toydiagram/resources/router/router.png
