+++ toydiagram/resources/switch/switch.png
